requirements.txt
streamlit
tensorflow
numpy
pillow
langchain
langchain-openai